fix: db

+++ server/server.js
@@ -58,11 +58,11 @@ if(process.env.NODE_ENV === 'production') {
 
   (async () => {
     try {
-      // const SQL = `SELECT * FROM USERS`;
-      // const [results, fields] = await db.promise().execute(SQL);
-      // if(results.length) return;
+      const SQL = `SELECT * FROM USERS`;
+      const [results, fields] = await db.promise().execute(SQL);
+      if(results.length) return;
+      console.log('delete table all...');
       await deleteAll();
-      await createAll();
     }
     catch(err) {
       console.error(err);
